push

+++ src/pages/api/delete/index.js
@@ -1,9 +1,29 @@
 import prisma from "../../../utils/prismaClient";
+import { getToken } from "next-auth/jwt";
 
 export default async function handle(req, res) {
+  const token = await getToken({ req });
+  const userEmail = token.email;
+
+  const user = await prisma.user.findUnique({
+    where: {
+      email: userEmail,
+    },
+  });
+
+  if (!user || !user.superAdmin) {
+    res
+      .status(403)
+      .json({ error: "Access denied. Only super admins can delete." });
+    return;
+  }
+
   if (req.method === "DELETE") {
     try {
       await prisma.inviteToken.deleteMany();
+      await prisma.installment.deleteMany();
+      await prisma.expenses.deleteMany();
+      await prisma.group.deleteMany();
       await prisma.weddingInvitationList.deleteMany();
       await prisma.wedding.deleteMany();
       await prisma.user.deleteMany();
